fix: dirfd validation, histogram overflow, and defensive bounds from code review
- Validate dirfd in aura_openat/aura_statx (reject negative except AT_FDCWD)
- Assert on buffer_pool_free with size=0 to catch caller bugs
- Bound backward-shift loop in buf_size_map_remove to prevent theoretical infinite loop
- Widen internal histogram total_count to uint64_t to prevent silent wraparound when adaptive is disabled; saturate to UINT32_MAX in public ABI struct

+++ engine/src/adaptive_engine.c
@@ -220,7 +220,7 @@ void adaptive_hist_record(adaptive_histogram_t *hist, int64_t latency_us) {
 }
 
 double adaptive_hist_p99(adaptive_histogram_t *hist) {
-    uint32_t total = atomic_load_explicit(&hist->total_count, memory_order_acquire);
+    uint64_t total = atomic_load_explicit(&hist->total_count, memory_order_acquire);
     if (total == 0) {
         return -1.0; /* No data */
     }
